changed to float

+++ SpaceSurvival/Assets/Scripts/Script/StatBar.cs
@@ -7,13 +7,13 @@ public class StatBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void setMax(int value)
+    public void setMax(float value)
     {
         slider.maxValue = value;
         slider.value = value;
     }
 
-    public void setValue(int value)
+    public void setValue(float value)
     {
         slider.value = value;
     }
